supplementary cycle: references to the interchangeable psalms
some of the Commons have a rubric saying that a certain psalm may be substituted in the
Midday Hour by another one in order to avoid repetition of the same psalm in the office of a single
day. References to these psalms are now under the 'normal' psalms to be substituted
in the supplementary cycle.