psalter: index of festal psalms: choose between instances of each psalm to avoid page-flipping
Up to now first occurrence of each psalm was referenced in the festal index,
which resulted to non-optimal suggestions.
e.g. for Good Friday's Laudes three distant places were suggested even if all the three
texts can be found in a single hour in the Psalter (Laudes, Friday of the 2nd week).

Now a simple, but in most cases successful algorithm is used to choose
which occurrence of a psalm present more than once in the Psalter to suggest for the particular
hour in order to eliminate unnecessary page-flipping and ribbons.

As a TODO remains to check if this algorithm is really ok as it is or if it results in any
harshly uncomfortable suggestion.