fix antiphon labels
these antiphons are not antiphons for Benedictus and Magnificat,
but antiphons sung alone in the commemoration after the ferial collect.

Interestingly Patric is the single saint in the Czech LotH having
"proper" commemoration antiphons. The other saints celebrated
in Lent are all equipped with commemoration antiphons taken from
the Benedictus/Magnificat antiphons of the corresponding Common.