Triduum booklet: on Easter Sunday always print "Deus in adiutorium"
since it's the first day the final Alleluia is being sung after the
whole season of Lent and aid is necessary to avoid confusion.

And yes, it's correct that "Deus in adiutorium" first appears at the
Midday Prayer: Office of Readings starts without any introduction with
the first reading, Lauds are, per a special rubric, on this day always
preceded by "Domine labia mea" and the invitatory.