add rubrics referencing the Commons
to all (and only) feasts having some proper chants and having such a
rubric ("fully ornated" formularies often miss a general rubric
referencing a Common and reference it with special rubrics concerning
e.g. the intercessions)

based on DMC diurnale 2007, but for reasons partly practical, partly
historical, rubrics for evangelists follow DMC diurnale 1994

The main reason for this is not having the rubrics printed in music
sheets (such a piece of information duplicating the breviary is not
particularly valuable there), but having the information in a
machine-readable format. It will be helpful once we will be generating
initial scaffolding/outline of the Antiphonale.