3 ant.s revised (back in 2015)
the variants in question have been marked for production ever since the
development file exists (i.e. since it was copied from the separate
branch 'variationes' to the directory of the same name in branch
'master'), but not applied to the production file.
They seem to have been first created in commit 0ee019c, 2013-05-03
(branch variationes, not merged in master and not intended to ever be
merged).

Only two days after that commit I wrote
http://www.inadiutorium.cz/blog/20150505-infrastruktura-projektu-in-adiutorium
and the updatefromvar.rb script and started using the workflow I've been
using ever since. But obviously it was too late for these three poor
antiphons, which stayed trapped in the mess of my original extremely
unpractical workflow up to this day.