(all) 3 ant.s revised
most of the work done on the day of the celebration and the days
immediately following, but back then I was undecided concerning the
Magnificat antiphon.
For this year there's no more time to spend on the formulary and it's
definitely better than it was before, so good enough, at least for now.