quality notices
In formularies which have two invitatory texts it's desirable, even if
the texts are similar, to set them to different melodies.
It's desirable to have choice not only in terms of words and ideas, but
also in terms of mode and style. That's what experience taught me
over the years. Commons of BVM and of Holy Men/Women have
always had different invitatory melodies because of structurally
dissimilar lyrics and I appreciate the kind of choice they offer.