checkcopies.rb: ?zacatek ignores differences in key signature
because in this corpus key signature is not considered a specification
of an important musical characteristic of the piece, but mere practical
or esthetical feature of scores containing many instances of bes.
It is legitimate for scores sharing the same beginning to differ in the
key signature (usually in the presence/absence thereof), the melody is
what really matters.