keep hour title in regular typeface
I'm not sure if the hour titles were ever actually set in small caps -
the markup command was there, but it took no effect, presumably because
the font didn't include the specified variant. Now it includes it, but
I got used to the title format as it has been for years now. Let it stay
that way. No small caps.