1 ant. revised
I just lost all variants collected yesterday
and, most importantly, the still unprocessed variants collected on
Pentecost and Holy Trinity, to an erroneous `git checkout` invocation.
This is the only change preserved, as it was already applied to the
production file and the checkout purged all unstaged changes from the
`variationes` directory (only).