+ 2 related antiphons
Quite probably there is the same Latin text behind both antiphons, but,
as usual, each religous institute translated its proper independently,
not knowing about the others.

This unexpected textual link between the offices of founders of
Salesians and Discalced Carmelites is particularly funny in Czech
Republic, as some of the first generation of post-1989 Czech Discalced
Carmelites are former Salesians.