replace a rubric with a literal copy of the referenced antiphon
we already manage two copies of the antiphon in question anyway and the
hard copy is more convenient for purposes of collecting content for the
mobile app, for the outline of the book etc.

The changed dependency tree takes advantage of automated child updating
(unlike a simple copy, ?-aleluja currently can't be handled automatically).