amend fial to pass auto-verification
'cast', unlike 'konec', doesn't have the smart handling of the
difference in divisio before alleluia. (And while it would be possible
to implement it, at least for the time being I consider it an overkill.
In the few cases where it would help it's just fine to use the cruder
'konec' check.)